cdi: move the CDI handling out of ateom-gvisor
containerd applies CDI to a pod's containers, and actor containers sit underneath
one, so whatever gives an actor a device reads the spec itself. Split by how far
it travels: internal/cdi parses a spec and resolves what a named set of devices
asks for, which any sandbox can use, while cmd/ateom-gvisor/internal/cdiinject
applies the result to a bundle on disk and stats the host for device numbers,
which only suits a sandbox on the host kernel.

The NVIDIA specifics -- which devices, which hooks may run, which library
directories -- are caller options rather than package globals. No behavior change,
and the tests run on any host rather than Linux only.
